LLM analysis implemented in the new arch Pipeline stages consolidated to one file Tested on 40 CFR Part 50
